# api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nodera Core API",
    "version": "0.1.0",
    "description": "Infrastructure, AI, Agent, and Operations control plane. This spec covers exactly the endpoints implemented today — see docs/ROADMAP.md in the repository for what's planned. Hand-maintained (docs/ROADMAP.md tracks generating it from code instead)."
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/api/v1/auth/signup": {
      "post": {
        "tags": ["auth"], "summary": "Create a user (no org membership yet)", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["email", "password", "display_name"],
          "properties": {
            "email": { "type": "string", "format": "email" },
            "password": { "type": "string", "minLength": 12, "description": "At least 12 characters, mixing letters with a number or symbol" },
            "display_name": { "type": "string" }
          }
        } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["auth"], "summary": "Authenticate and receive a session token", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["email", "password"],
          "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string" } }
        } } } },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": {
            "type": "object",
            "properties": {
              "session_token": { "type": "string" },
              "user": { "$ref": "#/components/schemas/User" },
              "organizations": { "type": "array", "items": { "$ref": "#/components/schemas/Organization" } }
            }
          } } } },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/api/v1/organizations": {
      "get": {
        "tags": ["organizations"], "summary": "List organizations the caller belongs to",
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": {
          "type": "array", "items": { "$ref": "#/components/schemas/Organization" }
        } } } } }
      },
      "post": {
        "tags": ["organizations"], "summary": "Create an organization; creator becomes owner",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["name", "slug"],
          "properties": { "name": { "type": "string" }, "slug": { "type": "string", "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$" } }
        } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Organization" } } } },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/organization": {
      "get": {
        "tags": ["organizations"], "summary": "Get the current organization",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Organization" } } } } }
      }
    },
    "/api/v1/infrastructure/nodes": {
      "get": {
        "tags": ["infrastructure"], "summary": "List nodes (infrastructure.read)",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }, { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Offset" }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NodePage" } } } } }
      },
      "post": {
        "tags": ["infrastructure"], "summary": "Register a node (infrastructure.manage)",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterNodeInput" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Node" } } } },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/infrastructure/nodes/{id}": {
      "get": {
        "tags": ["infrastructure"], "summary": "Get a node",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }, { "$ref": "#/components/parameters/PathId" }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Node" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/applications": {
      "get": {
        "tags": ["applications"], "summary": "List applications (applications.read)",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }, { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Offset" }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApplicationPage" } } } } }
      },
      "post": {
        "tags": ["applications"], "summary": "Register an application (applications.deploy)",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterApplicationInput" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Application" } } } },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/applications/{id}": {
      "get": {
        "tags": ["applications"], "summary": "Get an application",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }, { "$ref": "#/components/parameters/PathId" }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Application" } } } } }
      }
    },
    "/api/v1/api-tokens": {
      "get": {
        "tags": ["api-tokens"], "summary": "List the caller's own API tokens",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/APIToken" } } } } } }
      },
      "post": {
        "tags": ["api-tokens"], "summary": "Create an API token owned by the caller (organization.manage)",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAPITokenInput" } } } },
        "responses": {
          "201": { "description": "Created — raw token returned once", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatedAPIToken" } } } },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/api-tokens/{id}": {
      "delete": {
        "tags": ["api-tokens"], "summary": "Revoke one of the caller's own tokens",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }, { "$ref": "#/components/parameters/PathId" }],
        "responses": { "204": { "description": "No Content" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/organization/api-tokens": {
      "get": {
        "tags": ["api-tokens"], "summary": "List every non-revoked token in the org, any owner (organization.manage)",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/AdminAPIToken" } } } } } }
      }
    },
    "/api/v1/organization/api-tokens/{id}": {
      "delete": {
        "tags": ["api-tokens"], "summary": "Revoke any token in the org, regardless of owner (organization.manage)",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }, { "$ref": "#/components/parameters/PathId" }],
        "responses": { "204": { "description": "No Content" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/api/v1/service-accounts": {
      "get": {
        "tags": ["service-accounts"], "summary": "List service accounts (organization.manage)",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ServiceAccount" } } } } } }
      },
      "post": {
        "tags": ["service-accounts"], "summary": "Create a service account (organization.manage)",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["name"],
          "properties": { "name": { "type": "string" }, "description": { "type": "string" } }
        } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ServiceAccount" } } } } }
      }
    },
    "/api/v1/service-accounts/{id}": {
      "delete": {
        "tags": ["service-accounts"], "summary": "Disable a service account and revoke all its outstanding tokens (organization.manage)",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }, { "$ref": "#/components/parameters/PathId" }],
        "responses": { "204": { "description": "No Content" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
    "/api/v1/jobs/{id}": {
      "get": {
        "tags": ["jobs"], "summary": "Get a job",
        "parameters": [{ "$ref": "#/components/parameters/OrgHeader" }, { "$ref": "#/components/parameters/PathId" }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http", "scheme": "bearer",
        "description": "A session token (POST /auth/login) or an API token (POST /api-tokens / POST /service-accounts/{id}/api-tokens). See docs/API.md Authenticating."
      }
    },
    "parameters": {
...
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 200 } },
...
    },
    "responses": {
...
      "Conflict": { "description": "Conflicting state (e.g. duplicate key)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "RateLimited": { "description": "Too many requests", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" }, "email": { "type": "string" }, "display_name": { "type": "string" }
        }
      },
      "Organization": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" }, "name": { "type": "string" }, "slug": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}